Regression test for decoding 16x16 JPEG image with varying sampling factors
We don't decode this image to exactly the same pixel values as libjpeg. But I
have investigated the reason for the discrepancies and believe that both are
compliant with the spec. The spec does not require that every JPEG decoder
must always produce the exact same pixel values for the same compressed data;
the precision of decoders can vary.