[Bugfix] Imitate libjpeg's way of determining when to check for 'end of band'
Practically speaking, this means that we do not remember the previous
'end of band' for all MCUs in the same image component... but only for
previous scans of the _same_ MCU and same image component. It also means that
the 'end of band' for the first scan, which gives the high-order bits of each
coefficient, affects where we start checking for end of band in the first
successive refinement scan.

The JPEG spec isn't very clear about this.