[Bugfix] For progressive JPEGs, handle both interlaced and non-interlaced scans correctly
The same image may sometimes contain both interlaced and non-interlaced scans
(only if it is a progressive JPEG, though). For components with vertical
sampling factor > 1, the order of blocks will be different in interlaced
and non-interlaced scans. For the non-interlaced scans, blocks will run from
right to left, top to bottom (as usual). For the interlaced scans, they will
jump up and down between rows.

Also, the number of blocks in each MCU and the number of MCUs in the scan will
be different between the interlaced and non-interlaced scans. For the
non-interlaced scans, each MCU will just be a single block.

Our code for painting the pixel colors into a raster expects the samples to
be provided in the interlaced order. For both interlaced and non-interlaced
scans, ensure that the blocks of coefficients are stored in an array from
left to right, top to bottom (no jumping between rows). Then, when retrieving
those blocks of coefficients and preparing them for painting, retrieve them
in the interlaced order which the painting code expects.