[Bugfix] Still fixing bugs in calculation of block index (in coefficients array)
One problematic thing is that I am using the technical term 'MCU' in two
slightly incompatible ways; one is the sense used in the spec, which is some
number of blocks which are coded contiguously in a scan, and which is the unit
referred to by the 'restart interval'.

I am also using 'MCU' to refer to a group of blocks from each of the image
components, which, when taken together, fully cover some rectangular region
of pixels. (And are the minimal number of blocks which have that property,
given the varying sampling ratios of each image component.)